Struct futures_time::stream::Sample

pub struct Sample<S: Stream, I> { /* private fields */ }

Yield the last value received, if any, at each interval.

If no value was emitted during the last interval, no value is emitted and we skip to the next interval.

This struct is created by the sample method on StreamExt. See its documentation for more.